Fix build in C++0x mode
This fixes compiler errors (gcc 4.6 -std=gnu++0x on x86_64 Linux):

embedded/qwslock.cpp: In function `bool forceLock(int, int, int)':
embedded/qwslock.cpp:121:39: error: narrowing conversion of `semNum'
        from `int' to `short unsigned int' inside { } [-fpermissive]

(and equivalent errors in other lines/files)

See
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2011/n3242.pdf
Section 8.5.4/6

+++ src/gui/embedded/qwslock.cpp
@@ -121,7 +121,7 @@ QWSLock::~QWSLock()
     QWSSignalHandler::instance()->removeSemaphore(semId);
 }
 
-static bool forceLock(int semId, int semNum, int)
+static bool forceLock(int semId, unsigned short semNum, int)
 {
     int ret;
     do {
@@ -142,7 +142,7 @@ static bool forceLock(int semId, int semNum, int)
     return (ret != -1);
 }
 
-static bool up(int semId, int semNum)
+static bool up(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -155,7 +155,7 @@ static bool up(int semId, int semNum)
     return (ret != -1);
 }
 
-static bool down(int semId, int semNum)
+static bool down(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -168,7 +168,7 @@ static bool down(int semId, int semNum)
     return (ret != -1);
 }
 
-static int getValue(int semId, int semNum)
+static int getValue(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -217,10 +217,10 @@ void QWSLock::unlock(LockType type)
             return;
     }
 
-    const int semNum = type;
+    const unsigned short semNum = type;
     int ret;
     do {
-        sembuf sops = {semNum, 1, 0};
+        sembuf sops = { semNum, 1, 0 };
         if (semNum == QWSLock::BackingStore)
             sops.sem_flg |= SEM_UNDO;
